pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fcc-ad-tracker"
version = "0.1.0"
description = "Political Ad Revenue Estimator for Broadcast TV Credits using FCC OPIF data"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "anthropic>=0.77",
    "httpx>=0.28",
    "pydantic>=2.12",
    "pydantic-settings>=2.12",
    "typer>=0.15",
    "rich>=14.0",
    "PyMuPDF>=1.24",
    "pdfplumber>=0.11",
    "XlsxWriter>=3.2",
    "playwright>=1.48",
    "Pillow>=10.0",
]

[project.optional-dependencies]
ocr = [
    "paddleocr>=2.9",
    "paddlepaddle>=3.0",
    "pytesseract>=0.3",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.22",
]

[project.scripts]
fcc-tracker = "fcc_ad_tracker.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/fcc_ad_tracker"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
